requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==0.28.1
schedule==1.2.1
markdown==3.5.2
html2text==2020.1.16
PyYAML==6.0.1
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
python-dateutil==2.8.2
feedparser==6.0.11 